Parallel execution: Handle NetworkError exceptins in child process
Python Exceptions thrown in the child process are obviously handled
by the multiprocessing lib that is showing a traceback. The handler
of the NetworkError exception in tasks:execute was not receiving the
exception

Since we don't want to see the trace we are handling the exception
in the child process, showing an error.

For the reason of backwards compatibility env.use_exception_for['network']
is evaluated in the same manner as in tasks:execute.